Refactor test to reflect refactoring of SyscoinMessageLibrary
