+++ test/superblocks.js
@@ -1,3 +1,5 @@
+const truffleAssert = require('truffle-assertions');
+
 const { TestHelper } = require('@openzeppelin/cli');
 const { Contracts, ZWeb3 } = require('@openzeppelin/upgrades');
 
@@ -11,9 +13,6 @@ const SyscoinSuperblocks = Contracts.getFromLocal('SyscoinSuperblocks');
 
 contract('SyscoinSuperblocks', (accounts) => {
   let superblocks;
-
-  // raw syscon mind TX
-  // 067400000001013532f20fad408fb5beadaf4e2b4afbd86d12178d6a51ad4cec2ebc0f2fce93d20100000000feffffff020000000000000000fd2e076a4d2a07020150fd9e02f9029bf851a045a99ae455983d52604ee2a850ef7548b74ea118f1012297076dbb04b5348a0780808080808080a029cff91398b0056df3adfaa20bb49d3b416a0e69ee2c3117491ab36eca8465788080808080808080f8f180a082c0e775c7a55cd7dc0887cb15c90d760d994cd4d21fe26471164a9a1cc81161a034586b91f0fbbadaf00a6b3ffeeca8500255f6e8697f522a4c3c15c0f506c31ba0efdaeb23d58fcb4a52516e6856cb5cd9b19a5540cff4013a8f1e20b20ca51f5da00edb363b44a099c43f07c57083e05c5014ff4cbf2ea6ded88c009f12c8d0795ba0c3cdf1761eb751ff912eebb22b785e66976c159c193d7a4eb509e37d98c1bfc0a0595feaf2bbd299d9608f6c8533f9fbf0c34542ae423b35b8e89c432ea7645d61a07193d4763528c975c662910c5ca38424794bc8458e5102d8910f87e3edcbfab3808080808080808080f9015220b9014ef9014b8204ce843b9aca008307a1209438945d8004cf4671c45686853452a6510812117c80b8e45f959b690000000000000000000000000000000000000000000000000000000008f0d180000000000000000000000000000000000000000000000000000000002b7d3c01000000000000000000000000fbf4411309e690a6209b7f0d70ea304f8b40ac20000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000015002d68d3fe66b558010a3fa3ebf424bf1b8a59102400000000000000000000002ba01d91266755bdfe30c3b53a485afae06843bf21fb307b9f1947463002ce991f6ca06bdd1b6acc6016bf71e7ad4012e4e38d6552169e5d7148dbf0deca014794fd9e21a0f7b6f97b29bb3b107a1d1378f46f00750c4dacaa479369e4d599273ded5d98cd0105020150fd1304f90410f851a087ce28043d98064d651ec747b0ad1424857abda5ebe33cb512394893d0dacaa680808080808080a07965dbc2e42fbbaa7e6b24a3b93068513bea5a0c43a558dddefc2ae33b52f2898080808080808080f8f180a042f337f63d327f55f6c1498506df6bef6cc468107f163fcf9cb20d09a710ecf5a0991789340b01cd84de5499517a62d90a59eb0e44fe076be58b06704d87e0fbfaa0bd5a74c21752411a4442d5bbe4ffd0ecddcf3dee4a9b6166ee8e22b27fce0c7fa0ddf71b73d27d7bd2c06a734dda7505d590151177142905bf1169aed9cb09121da00b46675c3124c8235ba46d8d946f489f9e7e370b586edd3fdfa155cba6da02b0a0bad056ccfa51a0e98c05a8eaee60400170583e31b257bcf84b6db65db0a23e82a0ff2dcf5e1ec682c011fc64bce7e27b78061f2f899b2af4e8b3b9e0ef2ce2409b808080808080808080f902c720b902c3f902c0018305244eb9010000000000000000000000000000000000000000000000000000000000000100020000000000000000010000000000000000000000000000000000000000200000000000000000080001000008000000000000000000000000340000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000020000000000000000000000000000000000000000008120000000000000010000000002000000000000000000000000000000800000000000000000000000800010000000000000000000000000000000000000000000000000000000000000f901b5f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000008f0d180f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000000000000f8799438945d8004cf4671c45686853452a6510812117ce1a07def9f73ac6a7ac8fe4dc55f5257d48aed7e3f9d5247b0020598b87a5c369d82b840000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c1440530000000000000000000000000000000000000000000000000000000008f0d18021a06759019d118f4f19694ee8ed58167d2f460834bbb7dac5a0c5fd8ade3816e66f00a8ff530032401b060014ced26f5a296f082540c8f9cffee8015ecef6ad2c80d1f00800000000d40a9a3b00000000160014ced26f5a296f082540c8f9cffee8015ecef6ad2c02473044022034f3e994ff3c507217111664139236ec43a1b92895841cfffe91036afdb3018c022008bc2875f297bd9b2afa659db87af3b75a2a4d1f1845bc974b57c2e0b41f279a0121033e295585b49b152f0bdc61f10c8bfb6695d87e9eb7b1e1120ea2660b5d086cff00000000
 
   const randomAddress = accounts[0];
   const proxyAdmin = accounts[9];
@@ -238,6 +237,142 @@ contract('SyscoinSuperblocks', (accounts) => {
 
       result = await superblocks.methods.invalidate(id3, claimManager).send({ from: claimManager, gas: 300000 });
       assert.equal(result.events.InvalidSuperblock.event, 'InvalidSuperblock', 'Superblock invalidated');
+    });
+  });
+  describe("Test pure/view computation functions", () => {
+    it("bytesToBytes32", async () => {
+      const result = await superblocks.methods.bytesToBytes32("0x0102030405060708090a0b0c0d0e0f000102030405060708090a0b0c0d0e0f00", 0).call();
+      assert.equal(result, "0x0102030405060708090a0b0c0d0e0f000102030405060708090a0b0c0d0e0f00", "converted bytes are not the expected ones");
+    });
+    it("bytesToUint32", async () => {
+      const result = await superblocks.methods.bytesToUint32("0x01020304", 0).call();
+      assert.equal(result, 16909060, "converted bytes are not the expected ones");
+    });
+    it("bytesToUint16", async () => {
+      const result = await superblocks.methods.bytesToUint16("0x0150", 0).call();
+      assert.equal(result, 336, "converted bytes are not the expected ones");
+    });
+    it("getEthReceipt position encoded value field", async () => {
+      const receiptValue = "0xf902c0018305244eb9010000000000000000000000000000000000000000000000000000000000000100020000000000000000010000000000000000000000000000000000000000200000000000000000080001000008000000000000000000000000340000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000020000000000000000000000000000000000000000008120000000000000010000000002000000000000000000000000000000800000000000000000000000800010000000000000000000000000000000000000000000000000000000000000f901b5f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000008f0d180f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000000000000f8799438945d8004cf4671c45686853452a6510812117ce1a07def9f73ac6a7ac8fe4dc55f5257d48aed7e3f9d5247b0020598b87a5c369d82b840000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c1440530000000000000000000000000000000000000000000000000000000008f0d180";
+      const txBytes = "0x067400000001013532f20fad408fb5beadaf4e2b4afbd86d12178d6a51ad4cec2ebc0f2fce93d20100000000feffffff020000000000000000fd2e076a4d2a07020150fd9e02f9029bf851a045a99ae455983d52604ee2a850ef7548b74ea118f1012297076dbb04b5348a0780808080808080a029cff91398b0056df3adfaa20bb49d3b416a0e69ee2c3117491ab36eca8465788080808080808080f8f180a082c0e775c7a55cd7dc0887cb15c90d760d994cd4d21fe26471164a9a1cc81161a034586b91f0fbbadaf00a6b3ffeeca8500255f6e8697f522a4c3c15c0f506c31ba0efdaeb23d58fcb4a52516e6856cb5cd9b19a5540cff4013a8f1e20b20ca51f5da00edb363b44a099c43f07c57083e05c5014ff4cbf2ea6ded88c009f12c8d0795ba0c3cdf1761eb751ff912eebb22b785e66976c159c193d7a4eb509e37d98c1bfc0a0595feaf2bbd299d9608f6c8533f9fbf0c34542ae423b35b8e89c432ea7645d61a07193d4763528c975c662910c5ca38424794bc8458e5102d8910f87e3edcbfab3808080808080808080f9015220b9014ef9014b8204ce843b9aca008307a1209438945d8004cf4671c45686853452a6510812117c80b8e45f959b690000000000000000000000000000000000000000000000000000000008f0d180000000000000000000000000000000000000000000000000000000002b7d3c01000000000000000000000000fbf4411309e690a6209b7f0d70ea304f8b40ac20000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000015002d68d3fe66b558010a3fa3ebf424bf1b8a59102400000000000000000000002ba01d91266755bdfe30c3b53a485afae06843bf21fb307b9f1947463002ce991f6ca06bdd1b6acc6016bf71e7ad4012e4e38d6552169e5d7148dbf0deca014794fd9e21a0f7b6f97b29bb3b107a1d1378f46f00750c4dacaa479369e4d599273ded5d98cd0105020150fd1304f90410f851a087ce28043d98064d651ec747b0ad1424857abda5ebe33cb512394893d0dacaa680808080808080a07965dbc2e42fbbaa7e6b24a3b93068513bea5a0c43a558dddefc2ae33b52f2898080808080808080f8f180a042f337f63d327f55f6c1498506df6bef6cc468107f163fcf9cb20d09a710ecf5a0991789340b01cd84de5499517a62d90a59eb0e44fe076be58b06704d87e0fbfaa0bd5a74c21752411a4442d5bbe4ffd0ecddcf3dee4a9b6166ee8e22b27fce0c7fa0ddf71b73d27d7bd2c06a734dda7505d590151177142905bf1169aed9cb09121da00b46675c3124c8235ba46d8d946f489f9e7e370b586edd3fdfa155cba6da02b0a0bad056ccfa51a0e98c05a8eaee60400170583e31b257bcf84b6db65db0a23e82a0ff2dcf5e1ec682c011fc64bce7e27b78061f2f899b2af4e8b3b9e0ef2ce2409b808080808080808080f902c720b902c3f902c0018305244eb9010000000000000000000000000000000000000000000000000000000000000100020000000000000000010000000000000000000000000000000000000000200000000000000000080001000008000000000000000000000000340000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000020000000000000000000000000000000000000000008120000000000000010000000002000000000000000000000000000000800000000000000000000000800010000000000000000000000000000000000000000000000000000000000000f901b5f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000008f0d180f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000000000000f8799438945d8004cf4671c45686853452a6510812117ce1a07def9f73ac6a7ac8fe4dc55f5257d48aed7e3f9d5247b0020598b87a5c369d82b840000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c1440530000000000000000000000000000000000000000000000000000000008f0d18021a06759019d118f4f19694ee8ed58167d2f460834bbb7dac5a0c5fd8ade3816e66f00a8ff530032401b060014ced26f5a296f082540c8f9cffee8015ecef6ad2c80d1f00800000000d40a9a3b00000000160014ced26f5a296f082540c8f9cffee8015ecef6ad2c02473044022034f3e994ff3c507217111664139236ec43a1b92895841cfffe91036afdb3018c022008bc2875f297bd9b2afa659db87af3b75a2a4d1f1845bc974b57c2e0b41f279a0121033e295585b49b152f0bdc61f10c8bfb6695d87e9eb7b1e1120ea2660b5d086cff00000000";
+      const resultPos = await superblocks.methods.getOpReturnPos(txBytes, 4).call();
+      let pos = parseInt(resultPos);
+      pos += 3; // skip pushdata2 + 2 bytes for opreturn varint
+      
+      const result = await superblocks.methods.getEthReceipt(txBytes, pos).call();
+      assert.equal(result.toString(), receiptValue, "converted receipt bytes are not the expected ones");
+    });
+    it("getEthReceipt receipt value non-encoded", async () => {
+      const receiptValue = "0xf902c0018305244eb9010000000000000000000000000000000000000000000000000000000000000100020000000000000000010000000000000000000000000000000000000000200000000000000000080001000008000000000000000000000000340000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000020000000000000000000000000000000000000000008120000000000000010000000002000000000000000000000000000000800000000000000000000000800010000000000000000000000000000000000000000000000000000000000000f901b5f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000008f0d180f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000000000000f8799438945d8004cf4671c45686853452a6510812117ce1a07def9f73ac6a7ac8fe4dc55f5257d48aed7e3f9d5247b0020598b87a5c369d82b840000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c1440530000000000000000000000000000000000000000000000000000000008f0d180";
+      const txBytes = "0x067400000001010d0e68ec4449bddd7e91c5b1eb5a2c793705769dce5e295e77624b6c0f3e0e610100000000feffffff020000000000000000fd3f0b6a4d3b0bfd4e01f9014b8204ce843b9aca008307a1209438945d8004cf4671c45686853452a6510812117c80b8e45f959b690000000000000000000000000000000000000000000000000000000008f0d180000000000000000000000000000000000000000000000000000000002b7d3c01000000000000000000000000fbf4411309e690a6209b7f0d70ea304f8b40ac20000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000015002d68d3fe66b558010a3fa3ebf424bf1b8a59102400000000000000000000002ba01d91266755bdfe30c3b53a485afae06843bf21fb307b9f1947463002ce991f6ca06bdd1b6acc6016bf71e7ad4012e4e38d6552169e5d7148dbf0deca014794fd9efd9e02f9029bf851a045a99ae455983d52604ee2a850ef7548b74ea118f1012297076dbb04b5348a0780808080808080a029cff91398b0056df3adfaa20bb49d3b416a0e69ee2c3117491ab36eca8465788080808080808080f8f180a082c0e775c7a55cd7dc0887cb15c90d760d994cd4d21fe26471164a9a1cc81161a034586b91f0fbbadaf00a6b3ffeeca8500255f6e8697f522a4c3c15c0f506c31ba0efdaeb23d58fcb4a52516e6856cb5cd9b19a5540cff4013a8f1e20b20ca51f5da00edb363b44a099c43f07c57083e05c5014ff4cbf2ea6ded88c009f12c8d0795ba0c3cdf1761eb751ff912eebb22b785e66976c159c193d7a4eb509e37d98c1bfc0a0595feaf2bbd299d9608f6c8533f9fbf0c34542ae423b35b8e89c432ea7645d61a07193d4763528c975c662910c5ca38424794bc8458e5102d8910f87e3edcbfab3808080808080808080f9015220b9014ef9014b8204ce843b9aca008307a1209438945d8004cf4671c45686853452a6510812117c80b8e45f959b690000000000000000000000000000000000000000000000000000000008f0d180000000000000000000000000000000000000000000000000000000002b7d3c01000000000000000000000000fbf4411309e690a6209b7f0d70ea304f8b40ac20000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000015002d68d3fe66b558010a3fa3ebf424bf1b8a59102400000000000000000000002ba01d91266755bdfe30c3b53a485afae06843bf21fb307b9f1947463002ce991f6ca06bdd1b6acc6016bf71e7ad4012e4e38d6552169e5d7148dbf0deca014794fd9e21a0f7b6f97b29bb3b107a1d1378f46f00750c4dacaa479369e4d599273ded5d98cd0105fdc302f902c0018305244eb9010000000000000000000000000000000000000000000000000000000000000100020000000000000000010000000000000000000000000000000000000000200000000000000000080001000008000000000000000000000000340000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000020000000000000000000000000000000000000000008120000000000000010000000002000000000000000000000000000000800000000000000000000000800010000000000000000000000000000000000000000000000000000000000000f901b5f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000008f0d180f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000000000000f8799438945d8004cf4671c45686853452a6510812117ce1a07def9f73ac6a7ac8fe4dc55f5257d48aed7e3f9d5247b0020598b87a5c369d82b840000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c1440530000000000000000000000000000000000000000000000000000000008f0d180fd1304f90410f851a087ce28043d98064d651ec747b0ad1424857abda5ebe33cb512394893d0dacaa680808080808080a07965dbc2e42fbbaa7e6b24a3b93068513bea5a0c43a558dddefc2ae33b52f2898080808080808080f8f180a042f337f63d327f55f6c1498506df6bef6cc468107f163fcf9cb20d09a710ecf5a0991789340b01cd84de5499517a62d90a59eb0e44fe076be58b06704d87e0fbfaa0bd5a74c21752411a4442d5bbe4ffd0ecddcf3dee4a9b6166ee8e22b27fce0c7fa0ddf71b73d27d7bd2c06a734dda7505d590151177142905bf1169aed9cb09121da00b46675c3124c8235ba46d8d946f489f9e7e370b586edd3fdfa155cba6da02b0a0bad056ccfa51a0e98c05a8eaee60400170583e31b257bcf84b6db65db0a23e82a0ff2dcf5e1ec682c011fc64bce7e27b78061f2f899b2af4e8b3b9e0ef2ce2409b808080808080808080f902c720b902c3f902c0018305244eb9010000000000000000000000000000000000000000000000000000000000000100020000000000000000010000000000000000000000000000000000000000200000000000000000080001000008000000000000000000000000340000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000020000000000000000000000000000000000000000008120000000000000010000000002000000000000000000000000000000800000000000000000000000800010000000000000000000000000000000000000000000000000000000000000f901b5f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000008f0d180f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000000000000f8799438945d8004cf4671c45686853452a6510812117ce1a07def9f73ac6a7ac8fe4dc55f5257d48aed7e3f9d5247b0020598b87a5c369d82b840000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c1440530000000000000000000000000000000000000000000000000000000008f0d18021a06759019d118f4f19694ee8ed58167d2f460834bbb7dac5a0c5fd8ade3816e66f00a8ff53000c7f2e27001447314e644d681ba36d79e541ded7a73f074b265d80d1f0080000000080b9993b0000000016001447314e644d681ba36d79e541ded7a73f074b265d0247304402203ccfa7ffd7987dc9cf13214ec8e6fa5cc9ffb91e44d70a1a54654f25bf6ff0ba022039d595eb84ab6522a1784a20cdea1bf6adbb6444019af26dc1014479e371fcf701210348efe6431378340a7cbd3163527386be8071f96421ac333d7f4d5e01a27b476600000000";
+      const resultPos = await superblocks.methods.getOpReturnPos(txBytes, 4).call();
+      var pos = parseInt(resultPos);
+      pos += 3; // skip pushdata2 + 2 bytes for opreturn varint
+      
+      const result = await superblocks.methods.getEthReceipt(txBytes, pos).call();
+      assert.equal(result.toString(), receiptValue, "converted receipt bytes are not the expected ones");
+    });
+    it("getLogValuesForTopic", async () => {
+      const txBytes = "0x067400000001013532f20fad408fb5beadaf4e2b4afbd86d12178d6a51ad4cec2ebc0f2fce93d20100000000feffffff020000000000000000fd2e076a4d2a07020150fd9e02f9029bf851a045a99ae455983d52604ee2a850ef7548b74ea118f1012297076dbb04b5348a0780808080808080a029cff91398b0056df3adfaa20bb49d3b416a0e69ee2c3117491ab36eca8465788080808080808080f8f180a082c0e775c7a55cd7dc0887cb15c90d760d994cd4d21fe26471164a9a1cc81161a034586b91f0fbbadaf00a6b3ffeeca8500255f6e8697f522a4c3c15c0f506c31ba0efdaeb23d58fcb4a52516e6856cb5cd9b19a5540cff4013a8f1e20b20ca51f5da00edb363b44a099c43f07c57083e05c5014ff4cbf2ea6ded88c009f12c8d0795ba0c3cdf1761eb751ff912eebb22b785e66976c159c193d7a4eb509e37d98c1bfc0a0595feaf2bbd299d9608f6c8533f9fbf0c34542ae423b35b8e89c432ea7645d61a07193d4763528c975c662910c5ca38424794bc8458e5102d8910f87e3edcbfab3808080808080808080f9015220b9014ef9014b8204ce843b9aca008307a1209438945d8004cf4671c45686853452a6510812117c80b8e45f959b690000000000000000000000000000000000000000000000000000000008f0d180000000000000000000000000000000000000000000000000000000002b7d3c01000000000000000000000000fbf4411309e690a6209b7f0d70ea304f8b40ac20000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000015002d68d3fe66b558010a3fa3ebf424bf1b8a59102400000000000000000000002ba01d91266755bdfe30c3b53a485afae06843bf21fb307b9f1947463002ce991f6ca06bdd1b6acc6016bf71e7ad4012e4e38d6552169e5d7148dbf0deca014794fd9e21a0f7b6f97b29bb3b107a1d1378f46f00750c4dacaa479369e4d599273ded5d98cd0105020150fd1304f90410f851a087ce28043d98064d651ec747b0ad1424857abda5ebe33cb512394893d0dacaa680808080808080a07965dbc2e42fbbaa7e6b24a3b93068513bea5a0c43a558dddefc2ae33b52f2898080808080808080f8f180a042f337f63d327f55f6c1498506df6bef6cc468107f163fcf9cb20d09a710ecf5a0991789340b01cd84de5499517a62d90a59eb0e44fe076be58b06704d87e0fbfaa0bd5a74c21752411a4442d5bbe4ffd0ecddcf3dee4a9b6166ee8e22b27fce0c7fa0ddf71b73d27d7bd2c06a734dda7505d590151177142905bf1169aed9cb09121da00b46675c3124c8235ba46d8d946f489f9e7e370b586edd3fdfa155cba6da02b0a0bad056ccfa51a0e98c05a8eaee60400170583e31b257bcf84b6db65db0a23e82a0ff2dcf5e1ec682c011fc64bce7e27b78061f2f899b2af4e8b3b9e0ef2ce2409b808080808080808080f902c720b902c3f902c0018305244eb9010000000000000000000000000000000000000000000000000000000000000100020000000000000000010000000000000000000000000000000000000000200000000000000000080001000008000000000000000000000000340000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000020000000000000000000000000000000000000000008120000000000000010000000002000000000000000000000000000000800000000000000000000000800010000000000000000000000000000000000000000000000000000000000000f901b5f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000008f0d180f89b94fbf4411309e690a6209b7f0d70ea304f8b40ac20f863a08c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925a0000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c144053a000000000000000000000000038945d8004cf4671c45686853452a6510812117ca00000000000000000000000000000000000000000000000000000000000000000f8799438945d8004cf4671c45686853452a6510812117ce1a07def9f73ac6a7ac8fe4dc55f5257d48aed7e3f9d5247b0020598b87a5c369d82b840000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c1440530000000000000000000000000000000000000000000000000000000008f0d18021a06759019d118f4f19694ee8ed58167d2f460834bbb7dac5a0c5fd8ade3816e66f00a8ff530032401b060014ced26f5a296f082540c8f9cffee8015ecef6ad2c80d1f00800000000d40a9a3b00000000160014ced26f5a296f082540c8f9cffee8015ecef6ad2c02473044022034f3e994ff3c507217111664139236ec43a1b92895841cfffe91036afdb3018c022008bc2875f297bd9b2afa659db87af3b75a2a4d1f1845bc974b57c2e0b41f279a0121033e295585b49b152f0bdc61f10c8bfb6695d87e9eb7b1e1120ea2660b5d086cff00000000";
+      
+      // SHA3 of TokenFreeze(address,uint256,uint32)
+      const tokenFreezeTopic = "0xaabab1db49e504b5156edf3f99042aeecb9607a08f392589571cd49743aaba8d";
+
+      const expectedLogData = "0x000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c1440530000000000000000000000000000000000000000000000000000000008f0d180";
+
+      const resultPos = await superblocks.methods.getOpReturnPos(txBytes, 4).call();
+      let pos = parseInt(resultPos);
+      pos += 3; // skip pushdata2 + 2 bytes for opreturn varint
+
+      const ethTxReceipt = await superblocks.methods.getEthReceipt(txBytes, pos).call();
+      
+      const logData = await superblocks.methods.getLogValuesForTopic(ethTxReceipt, tokenFreezeTopic).call();
+
+      assert.equal(expectedLogData, logData, "Log data incorrect");
+    })
+
+    it("getBridgeTransactionId", async () => {
+      const logData = "0x000000000000000000000000b0ea8c9ee8aa87efd28a12de8c034f947c1440530000000000000000000000000000000000000000000000000000000008f0d180";
+      const expectedBridgeTransactionId = "150000000";
+
+      const bridgeTransactionId = await superblocks.methods.getBridgeTransactionId(logData).call();
+      assert.equal(expectedBridgeTransactionId, bridgeTransactionId.toString(), "Value incorrect");
+    })
+  })
+  describe('testParseTransaction', () => {
+    const keys = [
+      'L43bqxCXdZ1Lp6JkBoHE8pUQKi3BBgqrmBRYnNg4adZmRSRU534a',
+      'L4N2R2S2WRAnrdwnezs4kHaxsQkRNMrVwgGpHJNgJxGYT788LnGB',
+    ].map(utils.syscoinKeyPairFromWIF);
+    it('Parse simple transaction with only OP_RETURN and wrong version', async () => {
+      const tx = utils.buildSyscoinTransaction({
+        signer: keys[1],
+        inputs: [['edbbd164551c8961cf5f7f4b22d7a299dd418758b611b84c23770219e427df67', 0]],
+        outputs: [
+          ['OP_RETURN', 1000001, utils.ethAddressFromKeyPairRaw(keys[1])],
+        ],
+      });
+      tx.version = 0x01;
+      const txData = `0x${tx.toHex()}`;
+      const [ ret, amount, inputEthAddress, precision, assetGUID ] = Object.values(await superblocks.methods.parseBurnTx(txData).call());
+      assert.equal(ret, 10170, 'Error Parsing, wrong version');
+      assert.equal(amount, 0, 'Amount burned');
+      
+    });
+    it('Parse simple transaction with only OP_RETURN', async () => {
+      const tx = utils.buildSyscoinTransaction({
+        signer: keys[1],
+        inputs: [['edbbd164551c8961cf5f7f4b22d7a299dd418758b611b84c23770219e427df67', 0]],
+        outputs: [
+          ['OP_RETURN', 1000001, utils.ethAddressFromKeyPairRaw(keys[1])],
+        ],
+      });
+      tx.version = 0x7407;
+      const txData = `0x${tx.toHex()}`;
+      await truffleAssert.reverts(superblocks.methods.parseBurnTx(txData).send({from: randomAddress}));
+
+    });
+    
+    it('Parse simple transaction without OP_RETURN', async () => {
+      const tx = utils.buildSyscoinTransaction({
+        signer: keys[1],
+        inputs: [['edbbd164551c8961cf5f7f4b22d7a299dd418758b611b84c23770219e427df67', 0]],
+        outputs: [
+          [utils.syscoinAddressFromKeyPair(keys[1]), 1000001],
+          [utils.syscoinAddressFromKeyPair(keys[0]), 1000002],
+        ],
+      });
+      tx.version = 0x7407;
+      const txData = `0x${tx.toHex()}`;
+      await truffleAssert.reverts(superblocks.methods.parseBurnTx(txData).send({from: randomAddress}));
+    });  
+    it('Parse transaction with OP_RETURN in vout 1', async () => {
+      const tx = utils.buildSyscoinTransaction({
+        signer: keys[1],
+        inputs: [['edbbd164551c8961cf5f7f4b22d7a299dd418758b611b84c23770219e427df67', 0]],
+        outputs: [
+          [utils.syscoinAddressFromKeyPair(keys[0]), 1000002],
+          ['OP_RETURN', 1000001, utils.ethAddressFromKeyPairRaw(keys[1])],
+        ],
+      });
+      tx.version = 0x7407;
+      const txData = `0x${tx.toHex()}`;
+      await truffleAssert.reverts(superblocks.methods.parseBurnTx(txData).send({from: randomAddress}));
+      
+    });
+    it('Parse transaction with OP_RETURN', async () => {
+      const tx = utils.buildSyscoinTransaction({
+        signer: keys[1],
+        inputs: [['edbbd164551c8961cf5f7f4b22d7a299dd418758b611b84c23770219e427df67', 0]],
+        outputs: [
+          ['OP_RETURN', 1000001, utils.ethAddressFromKeyPairRaw(keys[1])],
+          [utils.syscoinAddressFromKeyPair(keys[0]), 1000002],
+        ],
+      });
+      tx.version = 0x7407;
+      const txData = `0x${tx.toHex()}`;
+      
+      await truffleAssert.reverts(superblocks.methods.parseBurnTx(txData).send({from: randomAddress}));
     });
   });
 });
